fix: repository test coverage 퍼센테이지 증대 및 불필요한 repository 정리
